Fullscreen works much better now, with two caveats.
1. HDR monitors such as mine automatically turn HDR off when entering fullscreen.

2. Without a graphics API (such as OpenGL) initialized, fullscreen causes the window to be invisible.

The latter seems like an unfortunate issue with Windows itself, while the former is probably because of some graphics API nonsense I have yet to understand. I have had a similar issue when playing DmC HD collection.

Instead of the "configuration" thing which existed before, now in app.init you pass whether you want to fullscreen to be auto, off, or on. For anyone having issues with fullscreen, for now simply forcing it off will have to do while I continue to investigate this.