set_size is too much of a shit show so I removed it.
I may add it again eventually. The issue is Windows puts the y-axis at the top left when I like it at the bottom left. So, in order to get set_size to work the way I would like, I would have to make it also move the window in such a way that it appears as if the bottom left is the y-axis.