Finished OpenGL 1.3.
I'm a little conflicted about what to do about the procedures that take a Texture__InternalFormat. I want to make as many procedures be type safe as possible. Unfortunately, Texture_InternalFormat has so many different formats that it seems impractical to make procedures that take it type safe. My compromise is to name the image/pixels parameter of that procedure have "raw" in the name to make it clear that the actual type of the pixels isn't byte.