Added text_input_state and set_text_input_flags. Changed text_input to make it return a cloned string. Removed gamepad events completely.
Originally, I thought gamepad events would be useful because you'd be able to do "keyboard input" from the gamepad by default. The problem is that the Unicode member of the KEYSTROKE struct is apparently unused, meaning I can't do that. So, at least for the use case I thought of, XInputGetKeystroke is completely useless to me. So, no gamepad events.