Removed Window_Mode_Free flags. Changed Fullscreen_Flags to a bool.
It might seem strange that I just removed a bunch of capabilities from Window_Mode_Free. Honestly, I just can't see myself using any of them, so I don't see why I should maintain them, since this is my library that I am making for my own purposes. Personally, I want to always know that the aspect ratio of the window will remain the same, so resizing is a no-no. Same thing with minimizing: it means the window all of a sudden has a width and height of 0, so really it has no defined aspect ratio. As for maximized, that's already a different window mode.

In video games, generally the way you change the resolution of the window is not by resizing the window itself. Instead, you go in the options menu and change the resolution there. It gives you a set of fixed aspect ratios to choose from, and that's it. If you want the game to be a maximized window, that's a whole option unto itself. That's what I'm modeling this all after.