I was just watching this podcast between Jon Blow and Thomas Brush: https://youtu.be/XhfZ1QEl-2s?si=kxfpAHI_VXeDFUgo
At some point Jon Blow talks about how hard it is to make games nowadays. He brings up fullscreen as an example. I want to focus on making this library as functional as possible. By functional, I don't mean like the programming paradigm. I mean that the things that matter should work, 100% of the time, one of those being fullscreen. I don't want there to be any fluff in the library at all. The use case is game development. That's it.

I just tested out topmost fullscreen with OpenGL, and AMD Software opened automatically. I take that as a win.